fix "famous quotations" fortune.
So, you know the "famous qoutations fortune?" The one that is this:
    " "
        -- Charlie Chaplin
    ...
When I was printing that qoute in-OS, I got this:
    pearlos> fortune
    " "
<U+2219> -- Charlie Chaplin
    ... this goes on ...
Imagine the "U+2219" as the actual ascii character U+2219. When I
accidentally slipped on my keyboard and hit the <TAB> key, you know
what came up:
    $ <U+2219>
I had accidentally used tabs instead of spaces.