update theming
When I was using pearlfetch I realized that the theme was called
"GREEN_ON_BLACK", but I was running the pascal theme! Then I
looked in the KSh, and i realized:

    "The theme is a variable, just that."

It's not anything that changes when the theme changes, like this
pseudocode here:

    int listen() {
        if (themeHasChanged) themeName = theme.Name;
        return 0;
    }

    int main() {
        while (true) {
            theme.Value = input(">");
            listen();
        }
        return 0;
    }

It's just a variable, like this pseudocode:

    int main() {
        themeName = "last default theme, but it has changed now.\n" \
                    "The problem is that I didn't get updated."
        while (true) {
            theme.Value = The_new_theme
            /**/
        }
    }

Meaning that just for a moment, the themes were wrong. I fixed it. But
then I thought:

    "Wait, theme-adding time!"

And then I added themes.