Add calculator
I decided to add a calculator to the OS to challenge my abilities of C!
That's what I was expecting... Instead, it turned into a nightmare of
type casting everywere and then that didn't work, so I had to write my
own caster, but we wouldn't have thi problem if C had competent type
casts! In a drem world this would do what I wanted it to:
    int number = (int)string;
But no! It had to be done another way! Well, it's done now.