Update panic interface
I felt it was time for a change. For too long now, this is what has
happened in pearlOS when you run "panic":

    % panic
    [PANIC] Debug panic.

That's a debug! It's been in this repo for multiple days. It was time
for a change. But I knew the normal kpanic() interface wasn't going to
cut it, kpanic() is a low level interface and the only reason why
advanced_cmds/ exists is because we need high level intefaces for the
Kernel SHell (KSH). We could obviously for example call the low level
/fs/fscore.h functions instead of using advanced_cmds/write_to_file.h,
read_file.h etc. But we don't. The Kernel SHell is a high level userland
wrapper for higher level interfaces. So, we simply add another
interface:

    .../pearlOS/kernel/advanced_cmds $ touch panic.h

panic.h is a higher level interface called by KSH. We strap together
/kernel/io.h, /kernel/kmsg.h and other lower level interfaces to create
this interface. This is all nicely packaged together in usrpanic(). We
can now update Kernel SHell to run usrpanic().