add type checking in calc
I was messing around in my operating system, and I decided to use the
calculator:

    $ calc
    num> invalid1
    num> invalid2
    op> +
    res> 0

Bullshit. It's bullshit to the power of fucking bullshit. But I knew
this could easily be solved, the first thing we need to do is to embed
type checking in the str_to_int() function inside of lib/conv.c:

    if (sizeof(number) != sizeof(int))
    {
        return -1;
    }

With this type embedding we can easily check type in
kernel/advanced_cmds/calc.h:

    if (str2int(s1) == -1 || str2int(s2) == -1)
    {
        /* ... */
        return -1
    }