Warningless GCC build [#2]
Patch the *tons* of memory bugs in kernel/mem.c. One of them was so
fucking stupid that I thought it was impossible to mess up so badly.

That memory bug was sizeof(x). My worst fucking enemy. What sizeof does
is that it returns the size of (bytes*)x. This is stupid. But then, it
also has a magical property. I hate magical properties.

This magical property is that if you take x (which is assumed to be a
list) divided by *x, like so:

    sizeof(x) / sizeof(*x)

It returns the list size! AKA, in an algebraic C format:

    _ x[] = {
        _ y[]
    }

    sizeof(x) / sizeof(*x) = strlen((_char)_join(for z in y do x))

This magical property is STUPID. Instead of using any fucking builtin,
like this (my ideal list size getter):

    #define len(a) ( \
        /* ... */ \
    )

    _ x[] = {
        _ y[]
    }

    len(x) /* length of _ x[] (=length of _ y[]) */

They use some division magic, which in reality doesn't make any fucking
sense. Like, how is ``(bytes*)x / (bytes*)*x`` = length of _ x[]!? It's
stupid.

The point is that I somehow missed this magical /sizeof(*x) when I was
doing a fucking for loop. This is basically the predicament I was in:

    _ x[] = {
        _ y[]
    }

    for (int i = 0; i < sizeof(x); i++) {
        x[i] = i
    }

Do you see the bug? If you do, good on you, because I didn't. These are
the changes to patch the bug:

    - for (int i = 0; i < sizeof(x); i++)
    + for (int i = 0; i < sizeof(x) / sizeof(*x); i++)

But hey! It's patched now!