patch type checking in calc
Since I am stupid, I initially checked type like this in lib/conv.c:

    if (sizeof(x) != sizeof(y))
    {
        kprints("{x:?} !type(y)");
    }

But then I realized, it was already converted. Assume that this is the
function:

    int str_to_int(char *x)
    {
        /* impl... */
        x = (int)x
        if (sizeof(x) != sizeof(int))
        {
            kprints("{x:?} !type(int)");
            return -1;
        }
        return x;
    }

    kprints(str_to_int("???"));

You'd expect this to return -1, right? Well, it doesn't. C automatically
converts all data into a smaller package, so something like this:

    .gcc
    char *x[] = "CHAR:X";
    int y = (int)x;

    .regi
    0x10 0x100
    0x20 0x50

Gcc truncates all of the bytes into a smaller package, which makes a
cast like the former, return something like ```21897264``, as the data
just goes haywire. As you could see in the .regi section of the last
example (showing sizeof), we lost 50 bytes in the compression, that
makes gcc go insane and die.

This is a difficult thing to solve, considering that you can't really
compare types if the size truncates. But there is a tool that can get us
out of this predicament.

Char comparisons. I despise char comparisons. It goes against my
philosophy of clean code. But here we go. All ASCII characters have a
specific byte value, something like this:

    .gcc
    char a = 'a'
    char b = 'b'
    char zero = '0'

    .regi
    0x400 0x48
    0x500 0x49
    0x600 0x0

But C also allows us to do clean char comparisons of the byte values:

    char x = 'b';
    if (x > 'a')
    {
        kprints("{x:'b'} > 'a'");
    }

    char y = 'y';
    if (y < 'z')
    {
        kprints("{y:'y'} < 'z'");
    }

So, to check if a char is covertable to an int, refer to the following
macro:

    #define char_is_int(ch) ( \
        if ((ch >= 'a' && ch <= 'z') || (ch >= 'A' && ch <= 'Z')) -1 \
        else 0 \
    )

Using this strategy of char comparisons, we can check if char is
convertable to int, and if not, we give up.