redefine library handling
Before we get into this commit message, I'd like you to answer a simple
question. Which one of these include methods do you like best:

    #include "../../../../dir1/dir2/dir3/file.h"
    #include <file.h>

If you picked the latter, congrats! You're not clinically insane! So
basically, I just added this to CFLAGS in the Makefile:

    -I./lib

And then, magically, the include method goes from this:

    #include "../../lib/io.h"

To:

    #include <io.h>

What we're doing in the Makefile is that we're basically telling the gcc
compiler to treat the ./lib directory the same as the /usr/include
directory. With this, code is cleaner.