Update Kernel SHell allocator
A few weeks ago I added a simple allocator to Kernel SHell, it works
like this:

    int allocator()
    {
        kmalloc(1)
        return 0;
    }

But this usage is very limited, as you can only allocate one byte per
call.

But now, it returns! The memalloc function! It takes the user input and
allocates it, basically:

    kmalloc(str_to_int(kinputs(), sizeof(kinputs) / sizeof(*kinputs)))

So yeah, fixed.