Re-add mk/iso directory
I realized that (sadly) I'm not always right. In an earlier commit I
decided to be stupid and remove this line in ./configure:

    mkdir -p mk/iso

Why did I do this? Well, I found it empty one time and thought "Well,
fuck it! if it's empty than why not remove it?" and removed it. But
then, I was sitting at my computer with my mediocre cup of coffee,
looking at the Makefile. And you know what fucking lines were in there!?

    cp $(.DEFAULT_GOAL) mk/iso/kernel.bin
    mkisofs -b kernel.bin -o dist/pearl.iso mk/iso/

Oh well. Turns out it was a util for ``make dist/pearl.iso``. Fuck it,
I'm becoming a farmer.