Patch str_to_int() function
I was sick today... And therefore not in the mood to code, but I decided
to do it anyways! That's when I learned that code generated by having a
migraine isn't that good.

What happened? Well, I tried to convert the str_to_int() function to a
function that would inhibit both an int and a bool, something like:

    (int|bool) calc()
    {
        if CONDITION
        {
            return BOOLEAN;
        }
        return INTEGER;
    }

But quickly realized that it was impossible. If you're new to the C
space you might not know that a boolean is just (1|0) in disguise,
meaning that I left a return 1; somewere in the code.

The problem is that... 1 is actually a pretty common number, which gets
even worse because it's always a check on the first element, so things
like this:

    pearl [/] > calc
    num> 10
    num> 20
    op> +

Don't work. This can be solved by simple defining an INVALID_INT macro,
which the INF macro is pretty good at because (int)~(1<<31) != trueint
as it is infinite. So this is the macro:

    #define INVALID_INT (int)INF

And you can then compare and check it.