feat: Switch license to the Hippocratic 3.0 Core License
As a socialist, I want to do something to better the world from things
like fascism, so I use the ANTI-FASCIST-MIT-LICENSE by jamiebuild (great
license btw)...

And then I wallow in my own sorrows listening to Penelope Scott and
Femtanyl for the rest of my life until I stumble upon the HL3
(Hippocratic License 3.0) license, and that changes everything. I *want*
to use this, so I do. Simple as that.